requirements.txt
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.14.0
cycler==0.12.1
filelock==3.16.0
fonttools==4.61.1
frozendict==2.4.7
fsspec==2025.1.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.24.6
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
narwhals==2.14.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
safetensors==0.4.6
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.52.1
sympy==1.13.2
tenacity==9.1.2
toml==0.10.2
tokenizers==0.21.4
torch==2.1.2
torchvision==0.16.2
tornado==6.5.4
transformers==4.41.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.66
